requirements.txt
# Core Dependencies
pyyaml>=6.0.1
astor>=0.8.1
pylint>=3.0.0

# Development Tools
pytest>=7.4.0
black>=23.0.0
mypy>=1.5.0

# Optional Dependencies
# For advanced features:
# transformers>=4.35.0
# torch>=2.1.0
# openai>=1.0.0